move env_flag
